+++ packages/race-cli/src/reporters/html-reporter.ts
@@ -2,24 +2,7 @@ import {LighthouseResultsWrapper} from '@racepoint/shared';
 import fs from 'fs/promises';
 import {BaseRacepointReporter} from '../types';
 import logger from '../logger';
-
-/*
-  Helper function to format ISO date and full URL for the report HTML
-  Matches the output from Lighthouse
-*/
-const formatFilename = (url: string, date: string) => {
-  const title = url.replace(/(http(s)?:\/\/)|(\/.*){1}/g, '');
-  const formattedDate = new Date(date);
-  const y = formattedDate.getFullYear();
-  // Format date values (months, seconds, etc) to always be 2 digits
-  const mo = ('0' + (formattedDate.getMonth() + 1)).slice(-2);
-  const d = ('0' + formattedDate.getDate()).slice(-2);
-  const h = ('0' + formattedDate.getHours()).slice(-2);
-  const mi = ('0' + formattedDate.getMinutes()).slice(-2);
-  const s = ('0' + formattedDate.getSeconds()).slice(-2);
-
-  return `${title}_${y}-${mo}-${d}_${h}-${mi}-${s}.race.html`;
-};
+import {formatFilename} from '../helpers';
 
 /**
  * Writes a single lighthouse result to an html file
@@ -32,10 +15,11 @@ export class HtmlReporter extends BaseRacepointReporter {
   }
 
   process = (results: LighthouseResultsWrapper): Promise<void> => {
-    const htmlPath = `${this._reportPath}/${formatFilename(
-      results.lhr.requestedUrl,
-      results.lhr.fetchTime
-    )}`;
+    const htmlPath = `${this._reportPath}/${formatFilename({
+      url: results.lhr.requestedUrl,
+      date: results.lhr.fetchTime,
+      suffix: 'race.html',
+    })}`;
 
     return fs
       .writeFile(htmlPath, results.report, {flag: 'w'})
